refactor: move Map type to public module
