--- src/internal/types.rs
@@ -1,2 +1,0 @@
-/// Map implementation used by the library
-pub type Map<K, V> = rustc_hash::FxHashMap<K, V>;
