+++ src/type_aliases.rs
@@ -2,10 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-//! Non exposed modules.
+//! Publicly exported type aliases.
 
-pub mod assignment;
-pub mod core;
-pub mod incompatibility;
-pub mod memory;
-pub mod partial_solution;
+/// Map implementation used by the library.
+pub type Map<K, V> = rustc_hash::FxHashMap<K, V>;
